change get_per_second parameters to long

+++ display.c
@@ -64,8 +64,8 @@ int spectrum_input(WINDOW *win, int c);
 /******************* HELPERS *******************/
 
 void
-get_per_second(unsigned int bytes, unsigned int duration,
-	       unsigned int packets, unsigned int retries,
+get_per_second(unsigned long bytes, unsigned long duration,
+	       unsigned long packets, unsigned long retries,
 	       int *bps, int *dps, int *pps, int *rps)
 {
 	static struct timeval last;
